Don't report test errors and failures twice.
Fixes #45.

The fix for #25 introduced this bug. The code to handle errors and failures would explicitly report the test failure to the nose TestResult, since the plugin otherwise tells nose to ignore these failures for flaky tests it wants to rerun.
This commit only does the explicit reporting if the test is marked flaky; otherwise, it lets nose handle the reporting.