Fixed an AttributeError in utils.ensure_unicode_string
If an object (like an Exception) is passed into ensure_unicode_string and the
exception has non-ASCII bytes in the message, it will raise a UnicodeError
(which will be caught) and then makes the false assumption that the argument
must be a string (and therefore have a .decode() method) resulting in an
AttributeError being raised.

This commit fixes the issue by wrapping it in a call to str() to ensure the
object's string representation is used.