Fix return values of nose plugin hooks.
Fixes #88

Nose plugin docs [0] say to return None from a hook to allow other
plugins to see the test. Flaky was returning False when it wanted
other plugins to see the test, which was causing incompatibility
with other plugins. This commit returns None from the addSuccess,
handleError, and handleFailure hooks where False would have been
returned (indicating that flaky was not handling the test).

[0]: http://nose.readthedocs.org/en/latest/plugins/interface.html#nose.plugins.base.IPluginInterface.addSuccess