WIP: print nice tracebacks in the flaky test report
Fixes #102 when you're using nose.

Breaks everything for pytest users, because the `err` argument, which
conains an (exc_type, exc_value, traceback) tuple, gets a the traceback
as a string.

I don't know if the right thing is to fix the caller to always pass the
same type (i.e. do the TB formatting in the nose plugin and always pass
a string here), or if this method needs to suppot polymorphism directly.